+++ app/controllers/exam_templates_controller.rb
@@ -3,7 +3,7 @@ class ExamTemplatesController < ApplicationController
   responders :flash, :http_cache
   respond_to :html
 
-  before_action :authorize_only_for_admin
+  before_action { authorize! }
 
   layout 'assignment_content'