+++ app/policies/annotation_category_policy.rb
@@ -1,0 +1,8 @@
+# Policy for annotation categories controller.
+class AnnotationCategoryPolicy < ApplicationPolicy
+  default_rule :manage?
+
+  def manage?
+    allowed_to?(:manage_assessments?, with: GraderPermissionPolicy)
+  end
+end