+++ app/policies/submission_policy.rb
@@ -1,4 +1,5 @@
 class SubmissionPolicy < ApplicationPolicy
+  alias_rule :manually_collect_and_begin_grading?, :collect_submissions?, :update_submissions?, to: :manage?
 
   def get_feedback_file?
     grouping = record.grouping
@@ -26,5 +27,13 @@ class SubmissionPolicy < ApplicationPolicy
 
   def before_release?
     !record.current_result.released_to_students
+  end
+
+  def manage?
+    allowed_to?(:manage_submissions?, with: GraderPermissionPolicy)
+  end
+
+  def can_run_tests?
+    allowed_to?(:run_tests?, with: GraderPermissionPolicy) && record.enable_test
   end
 end