+++ spec/controllers/submissions_controller_spec.rb
@@ -3,6 +3,59 @@ describe SubmissionsController do
     destroy_repos
   end
 
+  shared_examples 'An authorized admin and grader accessing #set_result_marking_state' do
+    context '#set_result_marking_state' do
+      let(:marking_state) { Result::MARKING_STATES[:complete] }
+      let(:released_to_students) { false }
+      let(:new_marking_state) { Result::MARKING_STATES[:incomplete] }
+      before :each do
+        @current_result = grouping.current_result
+        @current_result.update!(marking_state: marking_state, released_to_students: released_to_students)
+        post_as user, :set_result_marking_state, params: { assignment_id: @assignment.id,
+                                                           groupings: [grouping.id],
+                                                           marking_state: new_marking_state }
+        @current_result.reload
+      end
+      context 'when the marking state is complete' do
+        let(:new_marking_state) { Result::MARKING_STATES[:incomplete] }
+        it 'should be able to bulk set the marking state to incomplete' do
+          expect(@current_result.marking_state).to eq new_marking_state
+        end
+
+        it 'should be successful' do
+          expect(response).to have_http_status(:success)
+        end
+
+        context 'when the result is released' do
+          let(:released_to_students) { true }
+          it 'should not be able to bulk set the marking state to complete' do
+            expect(@current_result.marking_state).not_to eq new_marking_state
+          end
+
+          it 'should still respond as a success' do
+            expect(response).to have_http_status(:success)
+          end
+
+          it 'should flash an error messages' do
+            expect(flash[:error].size).to be 1
+          end
+        end
+      end
+
+      context 'when the marking state is incomplete' do
+        let(:marking_state) { Result::MARKING_STATES[:incomplete] }
+        let(:new_marking_state) { Result::MARKING_STATES[:complete] }
+        it 'should be able to bulk set the marking state to complete' do
+          expect(@current_result.marking_state).to eq new_marking_state
+        end
+
+        it 'should be successful' do
+          expect(response).to have_http_status(:success)
+        end
+      end
+    end
+  end
+
   describe 'A student working alone' do
     before(:each) do
       @group = create(:group)
@@ -242,7 +295,9 @@ describe SubmissionsController do
     end
   end
 
-  describe 'A TA' do
+  describe 'A grader' do
+    let(:grader) { create(:ta) }
+    let(:grader_permission) { grader.grader_permission }
     before(:each) do
       @group = create(:group)
       @assignment = create(:assignment)
@@ -263,15 +318,22 @@ describe SubmissionsController do
         repo.commit(txn)
 
         # Generate submission
-        Submission.generate_new_submission(Grouping.last,
+        submission = Submission.generate_new_submission(Grouping.last,
                                            repo.get_latest_revision)
+        result = submission.get_latest_result
+        result.marking_state = Result::MARKING_STATES[:complete]
+        result.save
+        submission.save
       end
-      @ta_membership = create(:ta_membership,
-                              grouping: @grouping)
+    end
+    context '#set_resulting_marking_state' do
+      let(:user) { create(:ta) }
+      let(:grouping) { @grouping1 }
+      include_examples 'An authorized admin and grader accessing #set_result_marking_state'
     end
     it 'should be able to access the repository browser.' do
       revision_identifier = Grouping.last.group.access_repo { |repo| repo.get_latest_revision.revision_identifier }
-      get_as @ta_membership.user,
+      get_as grader,
              :repo_browser,
              params: { assignment_id: @assignment.id, id: Grouping.last.id,
                        revision_identifier: revision_identifier,
@@ -281,7 +343,7 @@ describe SubmissionsController do
 
     it 'should render with the assignment_content layout' do
       revision_identifier = Grouping.last.group.access_repo { |repo| repo.get_latest_revision.revision_identifier }
-      get_as @ta_membership.user,
+      get_as grader,
              :repo_browser,
              params: { assignment_id: @assignment.id, id: Grouping.last.id,
                        revision_identifier: revision_identifier,
@@ -290,13 +352,79 @@ describe SubmissionsController do
     end
 
     it 'should be able to download the svn checkout commands' do
-      get_as @ta_membership.user, :download_repo_checkout_commands, params: { assignment_id: 1 }
+      get_as grader, :download_repo_checkout_commands, params: { assignment_id: 1 }
       is_expected.to respond_with(:missing)
     end
 
     it 'should be able to download the svn repository list' do
-      get_as @ta_membership.user, :download_repo_list, params: { assignment_id: 1 }
+      get_as grader, :download_repo_list, params: { assignment_id: 1 }
       is_expected.to respond_with(:missing)
+    end
+
+    let(:revision_identifier) do
+      @grouping.group.access_repo { |repo| repo.get_latest_revision.revision_identifier }
+    end
+
+    describe 'When grader is allowed to collect and update submissions' do
+      before do
+        grader_permission.manage_submissions = true
+        grader_permission.save
+      end
+      context '#collect_submissions' do
+        before do
+          post_as grader, :collect_submissions, params: { assignment_id: @assignment.id, groupings: [@grouping.id] }
+        end
+        it('should respond with 200') { expect(response.status).to eq 200 }
+      end
+      context '#manually_collect_and_begin_grading' do
+        before do
+          post_as grader, :manually_collect_and_begin_grading,
+                  params: { assignment_id: @assignment.id, id: @grouping.id,
+                            current_revision_identifier: revision_identifier }
+        end
+        it('should respond with 302') { expect(response.status).to eq 302 }
+      end
+      context '#update submissions' do
+        it 'should respond with 302' do
+          post_as grader,
+                  :update_submissions,
+                  params: { assignment_id: @assignment.id,
+                            groupings: [@grouping1.id],
+                            release_results: 'true' }
+          is_expected.to respond_with(:success)
+        end
+      end
+    end
+
+    describe 'When grader is not allowed to collect and update submissions' do
+      before do
+        grader_permission.manage_submissions = false
+        grader_permission.save
+      end
+      context '#collect_submissions' do
+        before do
+          post_as grader, :collect_submissions, params: { assignment_id: @assignment.id, groupings: [@grouping.id] }
+        end
+        it('should respond with 403') { expect(response.status).to eq 403 }
+      end
+      context '#manually_collect_and_begin_grading' do
+        before do
+          post_as grader, :manually_collect_and_begin_grading,
+                  params: { assignment_id: @assignment.id, id: @grouping.id,
+                            current_revision_identifier: revision_identifier }
+        end
+        it('should respond with 403') { expect(response.status).to eq 403 }
+      end
+      context '#update submissions' do
+        it 'should respond with 403' do
+          post_as grader,
+                  :update_submissions,
+                  params: { assignment_id: 1,
+                            groupings: ([] << @assignment.groupings).flatten,
+                            release_results: 'true' }
+          expect(response.status).to eq 403
+        end
+      end
     end
   end
 
@@ -361,55 +489,10 @@ describe SubmissionsController do
 
       around { |example| perform_enqueued_jobs(&example) }
 
-      context '#set_result_marking_state' do
-        let(:marking_state) { Result::MARKING_STATES[:complete] }
-        let(:released_to_students) { false }
-        let(:new_marking_state) { Result::MARKING_STATES[:incomplete] }
-        before :each do
-          @current_result = @grouping.current_result
-          @current_result.update!(marking_state: marking_state, released_to_students: released_to_students)
-          post_as @admin, :set_result_marking_state, params: { assignment_id: @assignment.id,
-                                                               groupings: [@grouping.id],
-                                                               marking_state: new_marking_state }
-          @current_result.reload
-        end
-        context 'when the marking state is complete' do
-          let(:new_marking_state) { Result::MARKING_STATES[:incomplete] }
-          it 'should be able to bulk set the marking state to incomplete' do
-            expect(@current_result.marking_state).to eq new_marking_state
-          end
-
-          it 'should be successful' do
-            expect(response).to have_http_status(:success)
-          end
-
-          context 'when the result is released' do
-            let(:released_to_students) { true }
-            it 'should not be able to bulk set the marking state to complete' do
-              expect(@current_result.marking_state).not_to eq new_marking_state
-            end
-
-            it 'should still respond as a success' do
-              expect(response).to have_http_status(:success)
-            end
-
-            it 'should flash an error messages' do
-              expect(flash[:error].size).to be 1
-            end
-          end
-        end
-
-        context 'when the marking state is incomplete' do
-          let(:marking_state) { Result::MARKING_STATES[:incomplete] }
-          let(:new_marking_state) { Result::MARKING_STATES[:complete] }
-          it 'should be able to bulk set the marking state to complete' do
-            expect(@current_result.marking_state).to eq new_marking_state
-          end
-
-          it 'should be successful' do
-            expect(response).to have_http_status(:success)
-          end
-        end
+      context '#set_resulting_marking_state' do
+        let(:user) { create(:ta) }
+        let(:grouping) { @grouping }
+        include_examples 'An authorized admin and grader accessing #set_result_marking_state'
       end
 
       context 'where a grouping does not have a previously collected submission' do