+++ spec/policies/submission_policy_spec.rb
@@ -87,4 +87,41 @@ describe SubmissionPolicy do
       end
     end
   end
+
+  describe '#can_run_tests' do
+    subject { described_class.new(assignment, user: user) }
+    context 'When the user is admin' do
+      let(:user) { create(:admin) }
+      let(:assignment) { build(:assignment, assignment_properties_attributes: { enable_test: true }) }
+      it { is_expected.to pass :can_run_tests? }
+    end
+    context 'When the user is grader and allowed to run tests' do
+      let(:user) { create(:ta, run_tests: true) }
+      let(:assignment) { build(:assignment, assignment_properties_attributes: { enable_test: true }) }
+      it { is_expected.to pass :can_run_tests? }
+    end
+    context 'When the user is grader and not allowed to run tests' do
+      # By default all the grader permissions are set to false
+      let(:user) { create(:ta) }
+      let(:assignment) { build(:assignment, assignment_properties_attributes: { enable_test: true }) }
+      it { is_expected.not_to pass :can_run_tests? }
+    end
+  end
+
+  describe '#collect and update submission' do
+    subject { described_class.new(user: user) }
+    context 'When the user is admin' do
+      let(:user) { create(:admin) }
+      it { is_expected.to pass :manage? }
+    end
+    context 'When the user is grader and allowed to collect and update submissions' do
+      let(:user) { create(:ta, manage_submissions: true) }
+      it { is_expected.to pass :manage? }
+    end
+    context 'When the user is grader and not allowed to collect and update submissions' do
+      # By default all the grader permissions are set to false
+      let(:user) { create(:ta) }
+      it { is_expected.not_to pass :manage? }
+    end
+  end
 end