+++ Changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 ## [unreleased]
 - Introduce the ability to designate criteria as 'bonus' marks (#4804)
+- Enable variable permissions for graders (#4601)
 
 ## [v1.10.0]
 - Issue #3670: Added API for adding and removing extra marks (#4499)